+++ src/client/components/map/Map.tsx
@@ -31,7 +31,6 @@ import {
   onlyUnlockedGeoUnits,
   getChildGeoUnits
 } from "./index";
-import AdvancedEditingModal from "./AdvancedEditingModal";
 import DefaultSelectionTool from "./DefaultSelectionTool";
 import MapMessage from "./MapMessage";
 import MapTooltip from "./MapTooltip";
@@ -356,7 +355,6 @@ const DistrictsMap = ({
     <Box ref={mapRef} sx={{ width: "100%", height: "100%", position: "relative" }}>
       <MapTooltip map={map || undefined} />
       <MapMessage map={map || undefined} maxZoom={maxZoom} />
-      <AdvancedEditingModal id={project.id} geoLevels={staticMetadata.geoLevelHierarchy} />
     </Box>
   );
 };